move DTO to separate files
