+++ novawallet/Modules/PayCard/Model/MercuryoTransferModel.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct MercuryoTransferModel {
+    let chainAsset: ChainAsset
+    let amount: Decimal
+    let address: AccountAddress
+}
