+++ novawallet/Modules/PayCard/PayCardInteractor.swift
@@ -1,29 +1,4 @@
-import BigInt
 import Foundation
-
-struct MercuryoTransferData: Decodable {
-    let id: String
-    let flowId: String
-    let amount: AmountDecimal
-    let currency: String
-    let network: String
-    let address: AccountAddress
-
-    enum CodingKeys: String, CodingKey {
-        case id
-        case flowId = "flow_id"
-        case amount
-        case currency
-        case network
-        case address
-    }
-}
-
-struct MercuryoTransferModel {
-    let chainAsset: ChainAsset
-    let amount: Decimal
-    let address: AccountAddress
-}
 
 final class PayCardInteractor {
     weak var presenter: PayCardInteractorOutputProtocol?
